refactor(actions): implement flux standard actions
All actions are now created by the same action creator `createAction`.
The middleware `forward` has been created in the `content` context,
it forwards to the `background` context the actions which have the
`forward` flag to `true` in their meta object.
It reduces greatly the boilerplate used to forward these actions as
well as insuring the separation of concerns.

Resolve #183 by implementing the proposal.