fix(service-message): ensure UI was opened before setting last shown date
Previous version was conceptually okay, though this one might failed less, there are still edge cases: for example if you repeatedly change tabs you might not see the UI was opened with a service message.
To avoid such cases one solution would be to rethink the main tab saga to be cancellable, if I switch to another tab I cancel the saga, because I'll relaunch it anyway when I'll go back there anyway. Also it would preserve a bit of cpu.