feat: Consider ToS have been accepted when the extension is installed.
This will prevent the consent window appearing too many times for some users..

Background behind this are the browser storage persistence issues. Without an account it's difficult keep track of the user consent.
We saw other extensions having the same kind of trouble for persisting user's config. Overall, extensions tend to consider user installation as consent, but for us it was really an issue since we don't store personal data anyway.

This implementation is rather naive, it just consider that ToS are accepted and open example page on first install.
This is on purpose, and will allow us to reactivate ToS easily or use part of it elsewhere.