fix: Browser action click not toggling UI
Since we remove the `activeTab` permission we lost access to the tab URL in some part of the API.
We have a little hack to get the tab URL thus it may not be such an issue.
However it's not very future proof and from my point of view (maybe I missed something) but the `activeTab` tab permission seems relevant in our case.

This implementation choice, assumes a tab may not have an URL.