chore: separate SENTRY_ENABLED and SENTRY_SEND_VERSION
We want to be able to generate a build with Sentry integrated and enabled without uploading the version to the Sentry platform.

`SENTRY_ORG`, `SENTRY_PROJECT`, `SENTRY_DSN` and `SENTRY_ENABLED` are now committed in staging and production .env file because they are no secret, and we know the values we want to use depending on the environment.

New var `SENTRY_SEND_VERSION` is not defined in any .env file but forced by npm script when using `buildVersion` instead of `build` script, thus triggering the upload of the built version to Sentry service. Such operation does need a .sentryclirc file with the user `token` to complete.